Dalsza refaktoryzacja. Zmienilem deklaracje metety simulate z klasy Simulator. Na wejsciu chce aby byly podawane dwie tablice. Pierwsza zawiera czasy zielonych swiatel, druga czerwonych. Upraszcza to integracje z algorytmem.

+++ BO/src/bo/project/logic/Intersection.java
@@ -21,11 +21,11 @@ public class Intersection extends Junction {
 		this.redLightTime = redLightTime;
 	}
 
-	public void setGreenLightTime(int newTime) {
+	public void setGreenLightTime(double newTime) {
 		this.greenLightTime = newTime;
 	}
 
-	public void setRedLightTime(int newTime) {
+	public void setRedLightTime(double newTime) {
 		this.redLightTime = newTime;
 	}
